feat: code refactoring.
