+++ src/main/java/ru/alfabank/api/exam/model/exception/BadTokenException.java
@@ -1,9 +1,9 @@
-package ru.alfabank.api.exam.util;
+package ru.alfabank.api.exam.model.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-@ResponseStatus(value = HttpStatus.BAD_REQUEST, reason = "Invalid currency token")
+@ResponseStatus(value = HttpStatus.BAD_REQUEST)
 public class BadTokenException extends RuntimeException {
 
     public BadTokenException(String message) {
