+++ src/main/java/ru/alfabank/api/exam/util/BadTokenExceptionHandler.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+import ru.alfabank.api.exam.model.exception.BadTokenException;
 
 @ControllerAdvice
 public class BadTokenExceptionHandler extends ResponseEntityExceptionHandler {
